Improvements + Spyalert changes
Voicemenu and Invis SpyAlert

+++ src/hacks/Backtrack.cpp
@@ -162,7 +162,7 @@ void Run()
         float distance, prev_distance_ticks = 9999;
 
         for (int i = 0; i < 12; ++i)
-        	sorted_ticks[i] = BestTickData{FLT_MAX, i};
+            sorted_ticks[i] = BestTickData{INT_MAX, i};
         for (int t = 0; t < ticks; ++t)
         {
         	if (headPositions[iBestTarget][t].tickcount)
